+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+- New features:
+  - Added the ability to enable manual command line snippet. [#40](https://github.com/Azure/PSDocs.Azure/issues/40)
+    - To enable parameter file snippet set configuration `AZURE_USE_COMMAND_LINE_SNIPPET`.
+    - See [about_PSDocs_Azure_Configuration] for details.
+
 ## v0.2.0-B2102005 (pre-release)
 
 What's changed since pre-release v0.2.0-B2101002: